HomeVC에서 RecentSummary 기능 분리
