+++ TLDR/TLDR/Presentaion/RecentSummary/View/RecentSummaryViewController.swift
@@ -5,14 +5,25 @@
 //  Created by 유정주 on 2023/05/05.
 //
 
-import Foundation
+import UIKit
 
-final class RecentSummaryViewController: BaseViewController<InfoView> {
-    
-    //MARK: - Life Cycles
+final class RecentSummaryViewController: UIViewController {
+
     override func viewDidLoad() {
         super.viewDidLoad()
-        
+
+        // Do any additional setup after loading the view.
     }
-    
+
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
 }
